Rename consumer to RabbitListener + remove useless imports
