+++ src/main/java/com/gramevapp/web/service/MQConfig.java
@@ -4,11 +4,8 @@ import org.springframework.amqp.core.*;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;
-import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter;
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
-import org.springframework.amqp.support.converter.MarshallingMessageConverter;
 import org.springframework.amqp.support.converter.MessageConverter;
-import org.springframework.amqp.support.converter.SimpleMessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -16,8 +13,7 @@ import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
-import org.springframework.oxm.Marshaller;
-import org.springframework.oxm.jaxb.Jaxb2Marshaller;
+
 
 @Configuration
 public class MQConfig {
@@ -48,9 +44,6 @@ public class MQConfig {
     @Bean
     public MessageConverter messageConverter() {
       return new Jackson2JsonMessageConverter();
-//        Jaxb2Marshaller marshaller = new Jaxb2Marshaller();
-//        marshaller.setPackagesToScan("com.gramevapp.web.service.RunnableExpGramEvWrapper");
-//        return new MarshallingMessageConverter(marshaller);
     }
 
     @Bean
